Saving Beyonders to save.yml

+++ src/main/java/de/firecreeper82/pathways/Sequence.java
@@ -46,7 +46,7 @@ public abstract class Sequence {
         return true;
     }
 
-    public void removeSpirituality(int remove) {
+    public void removeSpirituality(double remove) {
 
     }
